Fix submenu point calculations
Agent-Logs-Url: https://github.com/sagarchenchu/desktopAutomationHelper/sessions/db20ddbc-e6b5-4e21-b961-b255f2c0f49d

+++ src/DesktopAutomationDriver/Services/RecordingOverlayWindow.cs
@@ -5652,7 +5652,7 @@ public sealed class RecordingOverlayWindow : Form
                 Thread.Sleep(MenuNavigationDelayMs);
 
                 var right = new System.Drawing.Point(
-                    (int)Math.Round(rect.Right - Math.Max(8, Math.Min(20, rect.Width / 8))),
+                    (int)Math.Round((double)rect.Right - Math.Max(8.0, Math.Min(20.0, (double)rect.Width / 8.0))),
                     (int)Math.Round(rect.Top + rect.Height / 2.0));
 
                 if (TryPhysicalClickPoint(right, $"Open submenu {SafeElementName(item)}"))
